Address email comment #5 by DW.  Align horizontally home page nav icons.
5. Home page
There are four parts of this site and when I open the web page, two of these parts are not seen. I need to scroll down to find them. It seems to me that all of the material on this page should be presented without the need to scroll, there really isn't too much.  You shouldn't even need to scroll to see the AAA 'Show your card' message. Please reduce the size of the images and present them all on one line if possible. Since we're increasing the width of the page I'm hoping they all fit.

The four icons are now in a horizontal line across the page and have been moved up above the text.

'Get a Free Quote!' had to be shortened to 'Get a Quote' to fit without wrapping.