Address email comment #5.  Replace maps view with US Map.
Comment : The image of 'Where We Are' sorry to say looks like we are in the middle of the Gobi desert - not the impression we want to convey to our customers. If you need an image, please find a decent one of a US map and put that in there.

Actions : Cleaned up the home page view a little more.  Removed the 500px default height of pages. Replaced Google Map view icon with US map for 'Where We Are' icon.