Address email comment #6.  Anchor links on the What We Do page.
Comment : 6. What We Do
My understanding was that we were doing something that reflected parity with the existing site. To that extent I was expecting 2 sub-headers here 'Paintless Dent Repair' and 'Paint Repair'. If I click on Paint Repair I would go down to the section of the page talking about Paint Repair. Currently the Paint Repair details get lost at the bottom of the page. If we tell the user about it up at the top, they will at least know to click there for details. I also expected the same thing for the other headers - i.e. menu items at the top sending users the correct section of the page. You've done it correctly on the Who We Are page.

Action : Added named anchors to the top, and back to top links on the What We Do page.