Address comment #8 by DW.  Named anchor list of states on locations page.
8. Where We Are
Same comment as for #2 - we were expecting a list of states at the top that linked to the state listings below. What would be even better is if we have an image of the western states and click on the state to get to the listing below. Tell us if this is possible.

Action : Created a name anchor list of states that have locations that is generated dynamically.  Added back to top links in the body under each state.  Adding a clickable image map is not parity and was not in scope.