Address comment #10 by DW.  Google map in map view centered on Concord, CA.
10. Contact Us
Again, please don't have us out in the 'middle of the desert'. The Google map view is great, just make it 'map' instead of 'satellite', zoom out 4 clicks, and center it on Concord.

Action, Widened the map viewport so it fits new wider grid.  Changed map to Map view instead of Satellite view.  Centered on city center Concord, CA.  Caution, this may be misleading to users who use the tool to get directions, as now they are being directed to city center Concord, CA and not the business location.