Club DentPro members get a %75 discount after pricing the first dent (instead of normal %50).
Enhancement requested by Dave Whitgob via email on 1/24/2010.

DW:
Mitch brought to my attention that the formula in the Get A Quote does
not reflect the ClubDentPro 25% Members discount.  Not your fault, as
I did not put that in our original request.  The discount for anyone
getting a quote is correct, as it stands.  The first dent (largest) is
at the posted price, the second dent is 50% off.  The ClubDentPro member,
if the box is checked, will receive an additional 25% discount.