Change of requirements for applying Club DentPro discount.  Now %25 off entire estimate, not just dents.
The following reqs were provided by email from Sumit Guha on 1/25/2010:

1. Club Dentpro Discount - this discount is 25% off the total invoice, not just the discounted dents.
For example, I have a $200 dent, and 2 $100 dents. Without club dentpro checked, my quote is $300 ($200+($100+$100)/2). Now if club dentpro is checked, the bill should be $225 ($300*0.75).

Another point of clarification for bug #1:
25% off the total also applies to paint touch ups (which are not discounted for non-club members). So, for example, a $200 dent, 2 $100 dents and a $100 paint touch up would be quoted at $400 ($200+($100+$100)/2+$100) for non-club members, and $300 ($400*0.75) for club members.

I clarified with Sumit verbally that the Club DentPro %25 discount applies to ALL repairs and
that the discount is applied to what would have been the total price for a non-club member. There
are no excluded items from the discount.