New requirement that if there are 2x + dents of the same size, the second and 3rd also get the 50% discount.
Req documented in email from Sumit Guha on 1/25/2010:

2. The 50% off rule - the rule is the highest price dent is full price,
all others are 50% off (ignoring club dentpro). Now what happens when
there is more than one high price dent. The rule is that only one of
the high price dents is full price, others are discounted 50%.
For example, I have 2 $200 dents. The quote should be $300 ($200 + $200/2).
Currently the quote is coming back as $400.